# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo List Api",
        "description": "Demo Todo List Api",
        "contact": {
            "email": "joshuaamarfio1@gmail.com"
        },
        "version": "1.0.2"
    },
    "paths": {
...
        "/todo-api/public/api/v1/todos/{id}": {
            "get": {
                "tags": [
                    "Todo"
                ],
                "summary": "Get Detail Todo",
                "operationId": "42588b9579a572ea8473ac4252f6a97b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "example title"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "this is an example title"
                                                    },
                                                    "startdateTime": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2021-12-11T09:30:53.000000Z"
                                                    },
                                                    "enddateTime": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2021-12-11T10:30:53.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Todo"
                ],
                "summary": "Update Todo",
                "operationId": "7eb4232495ac46f4fc523e2ec1c311fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "startdateTime": {
                                                "type": "string"
                                            },
                                            "enddateTime": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "example title",
                                    "description": "example description",
                                    "startdateTime": "2022-12-19T12:06:00.000",
                                    "enddateTime": "2022-12-19T12:36:30.000"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "todo details updated"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "example title"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "this is an example title"
                                                    },
                                                    "startdateTime": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2021-12-11T09:30:53.000000Z"
                                                    },
                                                    "enddateTime": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2021-12-11T10:30:53.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}